[quantization] Rename _intrinsic to intrinsic
