+++ test/test_qat.py
@@ -5,7 +5,7 @@ from __future__ import unicode_literals
 
 import torch
 from torch.nn import Conv2d, BatchNorm2d, ReLU
-from torch.nn._intrinsic.qat import ConvBn2d, ConvBnReLU2d
+from torch.nn.intrinsic.qat import ConvBn2d, ConvBnReLU2d
 from torch.quantization.QConfig import default_qat_qconfig
 import torch.backends.mkldnn
 from common_utils import TestCase, run_tests
@@ -99,9 +99,9 @@ class IntrinsicQATModuleTest(TestCase):
             ).to(dtype=torch.double)
             qat_op.apply(torch.quantization.disable_fake_quant)
             if freeze_bn:
-                qat_op.apply(torch.nn._intrinsic.qat.freeze_bn_stats)
+                qat_op.apply(torch.nn.intrinsic.qat.freeze_bn_stats)
             else:
-                qat_op.apply(torch.nn._intrinsic.qat.update_bn_stats)
+                qat_op.apply(torch.nn.intrinsic.qat.update_bn_stats)
 
             # align inputs and internal parameters
             input = torch.randn(batch_size, input_channels, height, width, dtype=torch.double, requires_grad=True)
