+++ test/test_quantized_models.py
@@ -75,7 +75,7 @@ class ModelNumerics(QuantizationTestCase):
                 torch.quantization.prepare_qat(fq_model)
                 fq_model.eval()
                 fq_model.apply(torch.quantization.disable_fake_quant)
-                fq_model.apply(torch.nn._intrinsic.qat.freeze_bn_stats)
+                fq_model.apply(torch.nn.intrinsic.qat.freeze_bn_stats)
                 fq_model(calib_data)
                 fq_model.apply(torch.quantization.enable_fake_quant)
                 fq_model.apply(torch.quantization.disable_observer)
@@ -116,7 +116,7 @@ class ModelNumerics(QuantizationTestCase):
                     torch.quantization.prepare_qat(fq_model)
                     fq_model.eval()
                     fq_model.apply(torch.quantization.disable_fake_quant)
-                    fq_model.apply(torch.nn._intrinsic.qat.freeze_bn_stats)
+                    fq_model.apply(torch.nn.intrinsic.qat.freeze_bn_stats)
                     fq_model(calib_data)
                     fq_model.apply(torch.quantization.enable_fake_quant)
                     fq_model.apply(torch.quantization.disable_observer)
