+++ test/test_quantized_nn_mods.py
@@ -1,10 +1,10 @@
 import torch
 import torch.nn.quantized as nnq
 import torch.nn.quantized.dynamic as nnqd
-import torch.nn._intrinsic.quantized as nnq_fused
+import torch.nn.intrinsic.quantized as nnq_fused
 import torch.nn.quantized.functional as qF
 from torch.nn.quantized.modules import Conv2d
-from torch.nn._intrinsic.quantized import ConvReLU2d
+from torch.nn.intrinsic.quantized import ConvReLU2d
 import torch.quantization
 from common_utils import run_tests, IS_PPC, TEST_WITH_UBSAN
 from common_quantization import QuantizationTestCase, prepare_dynamic
@@ -228,7 +228,7 @@ class ModuleAPITest(QuantizationTestCase):
         qengine=st.sampled_from(("qnnpack", "fbgemm"))
     )
     def test_linear_api(self, batch_size, in_features, out_features, use_bias, use_fused, per_channel, qengine):
-        """test API functionality for nn.quantized.linear and nn._intrinsic.quantized.linear_relu"""
+        """test API functionality for nn.quantized.linear and nn.intrinsic.quantized.linear_relu"""
         if qengine not in torch.backends.quantized.supported_engines:
             return
         if qengine == 'qnnpack':
